audiolib: make mus_al_postamp a float and change volume calculation in AdLibDrv_MIDI_Service to FP math
Fixes terminx/eduke32#106, more or less.

+++ source/audiolib/include/al_midi.h
@@ -24,7 +24,7 @@ Foundation, Inc., 59 Temple Place - Suite 330, Boston, MA  02111-1307, USA.
 
 extern int AL_Stereo;
 extern int AL_AdditiveMode;
-extern int AL_PostAmp;
+extern float AL_PostAmp;
 
 typedef struct
 {
